Conflicts:
	cookbooks/wt_streaming_viz/CHANGELOG.md
	cookbooks/wt_streaming_viz/metadata.rb

Former-commit-id: a7ae4b65a331bdae0f239fe570e1e2658400cbfb [formerly d3f57771ded8b76c430420cab7b6bb10d926e517] [formerly b3402b93cfd964a6bda5aa9fde0f661318f4fff2] [formerly 9020b3b559d5ca9c7c8c9f3880997d5468b906a7 [formerly 90eb996cddc8f68c917f111ebe625240fa7fe0a5]] [formerly 0eef96968b3d885e0bc3003b0c790009b1395241 [formerly 27ba04db83b941a10cf69587bd9c260416b3ea2a] [formerly 8148cec6c563c43d98cdc11e41486121b833a6bd [formerly 252998726ef3751adacd074a60c27711288d22e1 [formerly aa95db2b005880bb971bcac95fa38a044da2edb9]]]]
Former-commit-id: 37e93b910d43405f6eb9251d8e8b3d4f658492ea [formerly 4b322fe740b6e37059ba1973e66f365d3b9f933c] [formerly 4f0e99aca7062ec790ecc11020d1a17fbf624bf0 [formerly 252998726ef3751adacd074a60c27711288d22e1]]
Former-commit-id: 04039095e8bb1937f83be1cf4ea0d12157b87e87 [formerly 621eef0203cb7d156deb1f7acd6957fe962c6c8b]
Former-commit-id: 72dad5662a2b56de5a050b12079cf57d3b8c2d47
Former-commit-id: eb043a52eefc81b649e79ec46ca661027f54d323

+++ cookbooks/java/recipes/openjdk.rb
@@ -26,6 +26,9 @@ pkgs = value_for_platform(
   ["centos","redhat","fedora","scientific","amazon"] => {
     "default" => ["java-1.#{jdk_version}.0-openjdk","java-1.#{jdk_version}.0-openjdk-devel"]
   },
+  ["debian","ubuntu"] => {
+    "default" => ["openjdk-#{jdk_version}-jdk","default-jre-headless"]
+  },
   ["arch","freebsd"] => {
     "default" => ["openjdk#{jdk_version}"]
   },
@@ -37,6 +40,7 @@ ruby_block  "set-env-java-home" do
   block do
     ENV["JAVA_HOME"] = java_home
   end
+  not_if { ENV["JAVA_HOME"] == java_home }
 end
 
 file "/etc/profile.d/jdk.sh" do
@@ -67,11 +71,10 @@ if platform?("ubuntu","debian","redhat","centos","fedora","scientific","amazon")
         Chef::Log.debug("glob is #{java_home_parent}/java*#{jdk_version}*openjdk*#{arch}")
         jdk_home = Dir.glob("#{java_home_parent}/java*#{jdk_version}*openjdk*#{arch}").first
         Chef::Log.debug("jdk_home is #{jdk_home}")
-        # delete the symlink if it already exists
-        if ::File.exists? java_home
-          FileUtils.rm_f java_home
+        if jdk_home
+          FileUtils.rm_f java_home if ::File.exists? java_home
+          FileUtils.ln_sf jdk_home, java_home
         end
-        FileUtils.ln_sf jdk_home, java_home
 
         cmd = Chef::ShellOut.new(
           %Q[ update-alternatives --install /usr/bin/java java #{java_home}/bin/java 1;
